requirements.txt
utils-nuuuwan
python-pptx
gTTS
pydub 